Create crossword.hs
Solution to Hackerrank challenge: https://www.hackerrank.com/challenges/crosswords-101/submissions/code/342530520

Input 11 strings: first 10 strings are the 10 rows of the crossword grid (+ for black squares and - for the empty squares) and then last string is a list of words separated by semicolons. Output is a the crossword grid with black squares filled in by the given words.